+++ CHANGELOG.md
@@ -19,6 +19,10 @@ Dates should be in`YYYY-MM-DD` format.
 - Fixed invalid HTML caused by markdown wrapped in paragraphs.
 - Moved the header and details into the nav bar, to persist while scrolling.
 
+### Fixed
+
+- Prevented non-authors from attempting to create a submission.
+
 ## 0.3.10 - 2026-04-05
 
 ### Added